refactor: change audio component structure
