+++ frontend/src/components/general/audio.tsx
@@ -1,0 +1,15 @@
+interface AudioProps {
+  musicPath?: string;
+}
+
+export const Audio = ({ musicPath }: AudioProps) => {
+  if (!musicPath) {
+    return null;
+  }
+
+  return (
+    <div className="flex justify-center">
+      <audio controls src={musicPath} />
+    </div>
+  );
+};
